fix: fix mock escape(query, body, curl, example)
Closes https://github.com/TongchengOpenSource/smart-doc/issues/680
